Fixed alarms and sleeping over two days
Whenever an alarm was set for the next day, and the alarm went off, the
current time would be later than the alarm time, but the alarm would
not be paused.  This meant that the time until the alarm goes off would
evaluate as a negative number, and the device would instantly wake up
when going into sleep.  This commit changes the code so that the timer
wakeup is only set if the alarm time is later than the current time.